Refactor model evaluation implementation
