+++ swan-pipeline/src/main/java/de/fraunhofer/iem/swan/SwanPipeline.java
@@ -52,7 +52,7 @@ public class SwanPipeline {
         featuresHandler.createFeatures();
 
         //Train and evaluate model for SRM and CWE categories
-        ModelEvaluator modelEvaluator = new ModelEvaluator(featuresHandler, options);
+        ModelEvaluator modelEvaluator = new ModelEvaluator(featuresHandler, options, testDataset.getMethods());
         modelEvaluator.trainModel();
 
         //TODO export final list to JSON file
