+++ swan-pipeline/src/main/java/de/fraunhofer/iem/swan/model/ModelEvaluator.java
@@ -1,26 +1,14 @@
 package de.fraunhofer.iem.swan.model;
 
 import de.fraunhofer.iem.swan.cli.SwanOptions;
+import de.fraunhofer.iem.swan.data.Method;
 import de.fraunhofer.iem.swan.features.FeaturesHandler;
+import de.fraunhofer.iem.swan.model.engine.MLPlan;
+import de.fraunhofer.iem.swan.model.engine.Meka;
+import de.fraunhofer.iem.swan.model.engine.Weka;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import weka.classifiers.Classifier;
-import weka.classifiers.bayes.BayesNet;
-import weka.classifiers.bayes.NaiveBayes;
-import weka.classifiers.evaluation.output.prediction.AbstractOutput;
-import weka.classifiers.functions.Logistic;
-import weka.classifiers.functions.SMO;
-import weka.classifiers.rules.JRip;
-import weka.classifiers.trees.DecisionStump;
-import weka.classifiers.trees.J48;
-import weka.core.Instances;
-import weka.filters.Filter;
-import weka.filters.MultiFilter;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.List;
+import java.util.Set;
 
 /**
  * Finds possible sources and sinks in a given set of system methods using a
@@ -43,11 +31,13 @@ public class ModelEvaluator {
 
     private FeaturesHandler features;
     private SwanOptions options;
+    private Set<Method> methods;
     private static final Logger logger = LoggerFactory.getLogger(ModelEvaluator.class);
 
     public ModelEvaluator(FeaturesHandler features, SwanOptions options, Set<Method> methods) {
         this.features = features;
         this.options = options;
+        this.methods = methods;
     }
 
     /**
@@ -59,86 +49,21 @@ public class ModelEvaluator {
 
         switch (Mode.valueOf(options.getLearningMode().toUpperCase())) {
 
-            case MANUAL:
-
-                //Phase 1: classify SRM classes
-                for (String srm : options.getSrmClasses())
-                    runManualEvaluation(features.getInstances().get(srm));
-
-                //Filter methods from CWE instances that were not classified
-                //into one of the SRM classes
-
-
-                //Phase 2: classify CWE classes
-                for (String cwe : options.getCweClasses())
-                    runManualEvaluation(features.getInstances().get(cwe));
-
-            case AUTOMATIC:
-                //return runAutomaticEvaluation(instances);
+            case MEKA:
+                logger.info("Evaluating model with MEKA");
+                Meka meka = new Meka(features, options, methods);
+                meka.trainModel();
+                break;
+            case WEKA:
+                logger.info("Evaluating model with WEKA");
+                Weka weka = new Weka(features, options);
+                weka.trainModel();
+                break;
+            case MLPLAN:
+                logger.info("Evaluating model with ML-PLAN");
+                MLPlan mlPlan = new MLPlan();
+                mlPlan.evaluateDataset(features.getInstances().get("train"));
+                break;
         }
-        return null;
-    }
-
-    /**
-     * Run AutoML training and evaluation on instances.
-     *
-     * @param instances list of instances
-     * @return
-     */
-    public HashMap<String, HashMap<String, String>> runAutomaticEvaluation(Instances instances) {
-
-        LinkedHashMap<String, HashMap<String, String>> fMeasure = new LinkedHashMap<>();
-
-        MLPlanExecutor mlPlanExecutor = new MLPlanExecutor();
-        // fMeasure.put("ML-Plan", mlPlanExecutor.evaluateDataset(instances));
-
-        //outputFMeasure(fMeasure);
-        return fMeasure;
-    }
-
-    /**
-     * @return
-     */
-    public HashMap<String, HashMap<String, String>> runManualEvaluation(Instances instances) {
-
-        LinkedHashMap<String, HashMap<String, String>> fMeasure = new LinkedHashMap<>();
-
-        List<Classifier> classifiers = new ArrayList<>();
-        classifiers.add(new BayesNet());
-        classifiers.add(new NaiveBayes());
-        classifiers.add(new J48());
-        classifiers.add(new SMO());
-        classifiers.add(new JRip());
-        classifiers.add(new DecisionStump());
-        classifiers.add(new Logistic());
-
-        //For each classifier, evaluate its performance on the instances
-        for (Classifier classifier : classifiers) {
-
-            MonteCarloValidator evaluator = new MonteCarloValidator();
-            evaluator.monteCarloValidate(instances, classifier, options.getTrainTestSplit(), options.getIterations());
-
-            for (String key : evaluator.getFMeasure().keySet())
-                logger.info("F-measure for {} using {}: {}", key, classifier.getClass().getSimpleName(), evaluator.getFMeasure().get(key));
-        }
-        return fMeasure;
-    }
-
-    /**
-     * Applies the Weka filters to the instances.
-     *
-     * @param instances instane set
-     * @param filters   array of filters
-     * @return instances with filter applied
-     */
-    public Instances applyFilter(Instances instances, MultiFilter filters) {
-
-        try {
-            filters.setInputFormat(instances);
-            return Filter.useFilter(instances, filters);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return null;
     }
 }
